third_party/blink/web_tests/fast/parser/document-write-partial-script-external-expected.txt

ALERT: PASS: This test passes if this string is alerted
